Support_Resistance_Evaluator/requirements.txt
pandas
numpy
streamlit
plotly
yfinance
requests
scipy
google-generativeai 
toml